+++ libraries/I2C_SCANNER/I2C_SCANNER.cpp
@@ -1,9 +1,10 @@
 //
 //    FILE: I2C_SCANNER.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.3.0
+// VERSION: 0.3.1
 //    DATE: 2022-08-29
 // PURPOSE: Arduino class to implement an I2C scanner.
+//     URL: https://github.com/RobTillaart/I2C_SCANNER
 
 
 #include "I2C_SCANNER.h"
@@ -158,13 +159,15 @@ uint8_t I2C_SCANNER::count(uint8_t start, uint8_t end)
 }
 
 
-bool I2C_SCANNER::setWireTimeout(uint32_t timeOut)
+bool I2C_SCANNER::setWireTimeout(uint32_t timeOut, bool reset)
 {
   if (_timeout != timeOut)
   {
     _timeout = timeOut;
     //  not all platforms support this.
-    //  _wire->setWireTimeout(timeOut);
+#if defined(WIRE_HAS_TIMEOUT)
+    _wire->setWireTimeout(timeOut, reset);
+#endif
     return true;
   }
   return false;
@@ -207,6 +210,30 @@ int I2C_SCANNER::_init()
   _wirePortCount++;
 #endif
   return _wirePortCount;
+}
+
+
+uint32_t I2C_SCANNER::getDeviceID(uint8_t address)
+{
+  uint8_t DEVICEID_ADDRESS = 0x7C;
+  _wire->beginTransmission(DEVICEID_ADDRESS);
+  _wire->write(address << 1);             //  address shifted one bit
+  int code = _wire->endTransmission(false);  //  explicit send a restart.
+  if (code != 0)
+  {
+    return 0xFF000000 + code;
+  }
+  if (_wire->requestFrom(DEVICEID_ADDRESS, (uint8_t)3) != 3)
+  {
+    return 0xFFFFFFFF;
+  }
+  //  build up return code
+  uint32_t deviceId = _wire->read();
+  deviceId <<= 8;
+  deviceId |= _wire->read();
+  deviceId <<= 8;
+  deviceId |= _wire->read();
+  return deviceId;
 }