+++ libraries/I2C_SCANNER/CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.3.1] - 2024-xx-xx
+## [0.3.1] - 2025-03-12
 - update examples
+- implement experimental **setWireTimeout()**
+- add experimental **uint32_t getDeviceID(uint8_t address)**
+- add experimental getSDA() + getSCL()
+- update readme.md
+- minor edits
 
 ## [0.3.0] - 2023-12-05
 - remove **begin(sda, scl)** to have less board specific dependencies.